+++ apps/pro-nextjs/src/app/(content)/workshops/[module]/[lesson]/(view)/exercise/page.tsx
@@ -1,5 +1,11 @@
+import * as React from 'react'
 import type { Metadata, ResolvingMetadata } from 'next'
-import { getLesson } from '@/lib/lessons-query'
+import { LessonProvider } from '@/app/(content)/tutorials/[module]/[lesson]/_components/lesson-context'
+import { ModuleProvider } from '@/app/(content)/tutorials/[module]/[lesson]/_components/module-context'
+import { LessonPage } from '@/app/(content)/workshops/[module]/[lesson]/(view)/shared-page'
+import { getExerciseSolution, getLesson } from '@/lib/lessons-query'
+import { getModuleProgressForUser } from '@/lib/progress'
+import { getWorkshop } from '@/lib/workshops-query'
 import { getOGImageUrlForResource } from '@/utils/get-og-image-url-for-resource'
 
 import Page from '../page'
@@ -29,7 +35,7 @@ export async function generateMetadata(
 	}
 }
 
-export default async function LessonPage({
+export default async function LessonExercisePage({
 	params,
 }: {
 	params: {
@@ -37,5 +43,20 @@ export default async function LessonPage({
 		lesson: string
 	}
 }) {
-	return <Page params={params} lessonPageType="exercise" />
+	const moduleLoader = getWorkshop(params.module)
+	const lessonLoader = getLesson(params.lesson)
+	const moduleProgressLoader = getModuleProgressForUser(params.module)
+
+	return (
+		<ModuleProvider moduleLoader={moduleLoader}>
+			<LessonProvider lessonLoader={lessonLoader}>
+				<LessonPage
+					lessonLoader={lessonLoader}
+					exerciseLoader={lessonLoader}
+					moduleLoader={moduleLoader}
+					moduleProgressLoader={moduleProgressLoader}
+				/>
+			</LessonProvider>
+		</ModuleProvider>
+	)
 }
